Incorrect Gradients in IE9
IE9 doesn't support the correct CSS gradients yet, so I didn't get to submit this to the 10k contest. It was fun attempting to do so, now I'm going to focus on extracting a game library, looking for other libraries to interact with or contribute to, and maybe port the code to a canvas based approach.